Got the internal of the new video upload working #1332
